Add 4 Widgets - Advanced SearchForm, Product Widget, Category Widget and User list
The Advanced Search form is a search form that allows your visitors to search with specific filter items.
The Product widget is an elegant widget that allows you to display the featured products of your store.
The Category widget is also an elegant widget that displays the most important categories of your store.
User List is a widget that allows you to create a user page with all the users of your website, you need to specify a category.